Слияние 3.3-extensions_HW и 4.1-jdbc
